// libs/features/medical-board/src/lib/shared/models/visiting-filter-request.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */
import { BilingualText } from '@gosi-ui/core';

export class VisitingFilterRequest {
  speciality: BilingualText[] = [];
  location: BilingualText[] = [];
  region: BilingualText[] = [];
}